Rename externalStore -> store
