+++ src/renderers/react/__tests__/store.spec.tsx
